New charts using dygraphs
We removed all D3.js from the chart, using dygraphs as our charting library.
We did so because it allows easier interactibe charts creation, implemented
as an overlay that appears as we click on one chart.

Also, we put Javascript code on monitor.js file and all styles on styles.css.
This should keep the code base cleaner.

TO-DO: Update charts dynamically.