Refatored divs IDs to allow focusing via parameter
We renamed the cells IDs to use the actual monitor IDs, allowing us to receive
a URL parameter named `id` with which we can focus on this given ID chart.

This will be useful when posting anomalies, because we will be able to give a
direct link to the anomalous chart.

We also made some changes to initially draw the plots while creating the table,
so the page looks more responsive now.